Move test_st_dwithin to test_predicates.py
